+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 ## Unreleased
 
 ### Changed
+- Ignore `DisallowedHost` exception in Sentry.
 - Upgrade all project dependencies.
 
 ### Fixed